pyproject.toml
[project]
name = "advent-of-code"
version = "0.3.0"
description = "My advent of code solutions"
authors = [{ name = "Souvlaki42", email = "souvlaki420@gmail.com" }]
license = "Unlicense"
urls = { repository = "https://github.com/Souvlaki42/advent-of-code" }
scripts = { aoc-manager = "main:cli" }
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["click>=8.3.0", "python-dotenv>=1.1.1"]

[tool.setuptools]
package-dir = { "" = "manager" }

[tool.uv]
package = true

[tool.setuptools.packages.find]
where = ["manager"]
